feat: added theme alert
