--- src/components/alertMessage/index.stories.mdx
@@ -1,19 +1,0 @@
-import { ArgsTable, Canvas, Meta, Story } from '@storybook/addon-docs';
-
-import AlertMessage from './index.tsx';
-
-<Meta title="Patterns/data display/Alert Message" />
-
-export const Template = (args) => <AlertMessage {...args} />;
-
-# Alert Message
-
-## Error
-
-<Canvas>
-  <Story name="Error" args={{ typeMessage: 'error' }}>
-    {Template.bind({})}
-  </Story>
-</Canvas>
-
-<ArgsTable story="Error" />
