+++ src/themes/components/alert.ts
@@ -1,10 +1,6 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 import { cssVar, mode } from '@chakra-ui/theme-tools';
-import {
-  PartsStyleFunction,
-  PartsStyleObject,
-  StyleFunctionProps,
-} from '@chakra-ui/styled-system';
+import { PartsStyleFunction, PartsStyleObject } from '@chakra-ui/styled-system';
 import { alertAnatomy as parts } from '@chakra-ui/anatomy';
 
 const $fg = cssVar('alert-fg');
@@ -13,46 +9,32 @@ const $bg = cssVar('alert-bg');
 const baseStyle: PartsStyleObject<typeof parts> = {
   container: {
     bg: $bg.reference,
-    px: '4',
-    py: '3',
+    padding: 'lg',
   },
   title: {
-    textStyle: 'heading1',
+    textStyle: 'heading4',
   },
   description: {
-    lineHeight: '6',
+    textStyle: 'body',
   },
   icon: {
     color: $fg.reference,
     flexShrink: 0,
-    marginEnd: '3',
-    w: '5',
-    h: '6',
-  },
-  spinner: {
-    color: $fg.reference,
-    flexShrink: 0,
-    marginEnd: '3',
-    w: '5',
-    h: '5',
+    marginEnd: 'lg',
+    w: 'sm',
+    h: 'sm',
   },
 };
 
-function getBg(props: StyleFunctionProps): string {
-  const { colorScheme: c } = props;
-  const lightBg = `${c}.100`;
-  const darkBg = `${c}.200`;
-  return mode(lightBg, darkBg)(props);
-}
-
 const variantLeftAccent: PartsStyleFunction<typeof parts> = (props) => {
   const { colorScheme: c } = props;
+  const bg = mode(`${c}.100`, `${c}.200`)(props);
   const fg = mode(`${c}.500`, `${c}.200`)(props);
   return {
     container: {
-      [$bg.variable]: getBg(props),
+      [$bg.variable]: `colors.${bg}`,
       [$fg.variable]: `colors.${fg}`,
-      paddingStart: '3',
+      paddingStart: 'lg',
       borderStartWidth: '4px',
       borderStartColor: $fg.reference,
     },
@@ -65,7 +47,6 @@ const variants = {
 
 const defaultProps = {
   variant: 'left-accent',
-  colorScheme: 'blue.100',
 };
 
 export default {
