+++ src/themes/components/index.ts
@@ -14,10 +14,10 @@ import Divider from './divider';
 import Checkbox from './checkbox';
 import Button from './button';
 import ArticleTeaser from './articleTeaser';
-import AlertMessage from './alertMessage';
+import Alert from './alert';
 
 export const components = {
-  AlertMessage,
+  Alert,
   ArticleTeaser,
   Button,
   Checkbox,
