+++ src/themes/shared/borders.ts
@@ -1,5 +1,7 @@
+/* eslint-disable @typescript-eslint/naming-convention */
+
 export const borders = {
   none: 0,
-  // eslint-disable-next-line @typescript-eslint/naming-convention
   '1px': '1px solid',
+  '4px': '4px solid',
 };
